Add tests proving that each column has to be proxied separately
Proxying the entire input data frame results in a flattened data frame. Even if we expand `condition` to match the new number of columns, this would result in incorrect comparisons. If `x` is a df-col and the comparison is `>`, we want to apply the `>` condition to each row as a whole, not to each column in the row separately. In particular, the first broken tie controls the comparison result for the entire row!